requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
PyYAML==5.4.1
pyaml==20.4.0
requests==2.25.1
forex-python==1.5
